Demonstrate bug in dataclass transform options `kw_only`
Summary:
In the midst of refactoring the dataswarm transform code to make more
sense and commenting how it works in more detail, I realized that our
`kw_only` support is completely broken because we mixed up the default
vs callsite-specific cases.

This commit adds test cases to illustrate the problem, which will be fixed
in an upcoming cleanup diff.

Reviewed By: grievejia

Differential Revision: D55981003

fbshipit-source-id: aaf1e627523c75b9cfea848d36b04674a7bd2333