Prepare for the split between generations and parameter sources
Summary:
# Context
There are two different kind of sources in models:
* What we will call **generations**: Those are sources that are propagated to callers of the model. They are usually inferred by the analysis.
* What we will call **parameter sources**: Those are sources that must be instantiated in the body of the modeled function. They are usually specified by users.

Here are examples of **generations**:
```
def foo():
  return source() # generation on port=return

def bar(l):
  l.append(source()) # generation on port=argument(l)

def baz():
  x = None

  def nested():
    nonlocal x
    x = source() # generation on port=captured_variable(x)
```

Users can specify **parameter sources** using models:
```
# in .pysa
def foobar(x: TaintSource[Test]): ...

# in .py
def foobar(x):
  sink(x) # issue since x is tainted

def foobar_caller():
  x = 0
  foobar(x)
  sink(x) # not an issue, parameter x is only tainted in the body of foobar
```

# Problem
Right now, generations and parameter sources are stored in the same data structure.
We simply assume that a source on a parameter is a parameter source, and that a source on the return port is a generation.
This means we cannot actually infer the right model for the function `bar` above.

For captured variables, we added a flag in the access path root (`CapturedVariable{name; generation_if_source}`) to distinguish between the two cases. While this work, this is less than ideal.

# Solution
Let's actually store parameter sources and generations in different components of the model.

# This diff
This diff prepares for the change mentioned above, by renaming `source_taint` in the model into `generations`, and introducing a `parameter_sources` component. For now, this is unused. The actual change will happen in the following diff.

Reviewed By: tianhan0

Differential Revision: D56014367

fbshipit-source-id: 9c0573b87572911415467c0bdb02dc987e7a8c9b