Rename confusing function
Summary:
In Pyre, the word "qualifier" almost always means a Reference.t,
usually one that we think corresponds to a python module name.

In the context of this helper function, that name doesn't make sense,
it's checking whether the outermost parametric type is `typing.Final`,
which is not checking anything about a "qualifier" in our usual
terminology.

Reviewed By: kinto0

Differential Revision: D56169440

fbshipit-source-id: 6d10b97a2b8cb549a35f22cff31af4bdde023009