Split test cases in assignTest.ml
Summary:
Converting to the usual style where each case is a separate
ounit test, runnable in isolation, and labeled with line numbers
so that any failures are easy to track down.

In this case I wanted to be able to use the tests to print-debug
as I figure out where exactly `Final` is handled in our codebase, which
is relevant to fixing dataclasses `frozen` behavior.

Reviewed By: grievejia

Differential Revision: D56168348

fbshipit-source-id: 7f222bb4517408be4543cc22593df3b188161aea