Document conformance failure in a test
Summary:
Pyre currently correctly marks fields of frozen dataclasses as
not updatable, but fails to complain if we try to inherit from a
non-frozen dataclass in a frozen dataclass (which the spec disallows;
frozen-ness is like read-only in that it behaves like a supertype
of normal mutable read-write access).

This was caught by the conformance tests, and I'm adding a test case
with a TODO illustrating the problem in our code.

Reviewed By: grievejia

Differential Revision: D56168343

fbshipit-source-id: fd0df3144546ee60ea05aafe3c754e08c203a25f