Infer generation sources on self
Summary:
# Context
Pysa does not currently propagate sources assigned to parameters of functions (including their attributes), including the implicit `self` parameter.
The reason is that it can lead to false positives and performance regressions in the analysis.

This means this flow would not be found:
```
class Foo:
  def mutates(self):
    self.bar = source()
  def issue(self):
    self.mutates()
    sink(self.bar) # Current false negative.
```

The only exception is constructors and property setters where we do infer sources assigned to attributes of self.

# This diff
This diff stack aims at inferring and propagating sources assigned to `self` attributes.

First, this diff implements the inference of those sources. More concretely, that means our models should now include generations for `self` attributes.
In a follow-up, we will actually propagate those.

Note that we still do NOT infer sources on other arguments.

Reviewed By: alexkassil

Differential Revision: D56065426

fbshipit-source-id: 8e858d70e69f5c8025518d84dd4f2b75b23a063c