Create model annotations when parsing decorators
Summary:
Instead of parsing decorators one per one and updating the final model, we can use the same strategy as for regular taint annotations: we parse the annotation, create a `ModelAnnotation.t`, and then instantiate all model annotations into a model at the end.
This makes refactorings easier. It would also allow model queries to create sanitizers in the future.

Reviewed By: alexkassil

Differential Revision: D56238687

fbshipit-source-id: f655b23429e078432e34468b0863544097e8b61d