Get nested configurations with glob
Summary:
On testing the per-target migration script locally, I noticed I wasn't able to migrate an example configuration.

It turns out that in the old logic, the `.pyre_configuration.local` getting migrated would be included in the `has_nested_configurations()` check.

I thought the logic I created earlier for finding nested configurations could be useful for other scripts, so I decide to turn this into a `get_nested_configuration_paths()` function so that the logic can be used in the per-target migration script too.

Reviewed By: MaggieMoss

Differential Revision: D56267667

fbshipit-source-id: 528aed7f47a61442577b471e3ff88b5b40c5e539