Placeholder stubs have to be stubs
Summary:
The current `legacy_resolve_exports` code understands that
if it fails to find a name it just gives up and returns that name,
without worrying about whether the name is defined. This behavior
implicitly treats *every* package as though lookups in nonexistent
modules are covered by an empty stub.

But `resolve_exports` explicitly understands placeholder stubs,
and importantly *only a stub file can be a placeholder stub*.

This test case was using a "placeholder stub" comment inside a
*normal* module. That's not valid, and the `Module.t` code that
powers `resolve_exports` knows that it's not valid.

Fixing this makes it more possible to cut over.

Reviewed By: grievejia

Differential Revision: D56310800

fbshipit-source-id: f7d3d33533aefb5002a59ee78cd6b98d13ea67fa