Add more detailed comments to `resolve_exports`
Summary:
This commit walks through `resolve_exports` at the level of detail
someone unfamiliar with the code would need - especially anyone
who hasn't memorized some of the less-obvious semantics at play
(the current code had some handy comments already, but I want a
perfect mental model and a lot of details were being elided about
exactly what semantics each block of code was dealing with).

I flipped the if / else blocks in one spot which is debatable:
sometimes people advocate for always using `if <condition>` rather
than `if not <condition>` as "clean code", but in this case the
condition is catching an edge case, and it's much easier to explain
what's happening in comments if we handle the "normal" case of
`from x import y` before the special case of `from . import y`.

Reviewed By: grievejia

Differential Revision: D56247768

fbshipit-source-id: 648f750cbcc2dace6354350fcff5a8ea08d6773a