Move the logic to compute the modeled callable in a single place
Summary:
Right now, the model parser logic to infer what the callable should be is spread all around the place.
For instance, we compute the callable directly from `parse_statement`, but fix it up later in `create_parsed_signature`.
Let's move all the logic to compute the callable in a single place.

Let's also stop using `Define.Signature.t` and only extract the members that we need.
We can use this opportunity to differentiate taint decorators (`Sanitize`, `SkipAnalysis`, etc.) and decorators on the original function (`property`, `property.setter`, etc.).

This refactoring will allow us to model top level functions in a follow-up.

Reviewed By: tianhan0

Differential Revision: D56315491

fbshipit-source-id: 0f13af8ccf545404b93968b65c9ffe962edfa295