Stop treating constructors, property setters and setitem as returning self
Summary:
# Context
We currently treat constructors (e.g, `__init__`), property setters and `__setitem__` methods as implicitly returning `self`.
This was convenient since it allowed us to re-use most of the tito logic.

Now, we can properly infer taint-in-taint-out to `self`, as well as sources assigned to self attributes. Hence, this trick is not necessary anymore. Let's remove it so our models are consistent.

Note that we will keep supporting user models on the return value of these models, for backward compatibility (see D56300979).

Reviewed By: tianhan0

Differential Revision: D56301056

fbshipit-source-id: f22de5f2804436ec9cf283a31e9a6822ff16694a