Respect PEP 561 in lazy module tracker
Summary:
This is a bugfix for lazy module tracking (which powers
the code-navigation server, including "global lazy" IDE
type errors).

The old code did not understand that within any given
search roots there are actually *two* potential
relative paths for the parent directory of a Python
module:
 - the one where the chunks of the path are the chunks
   of the (prefix of) the module name, e.g.
   for `foo.bar.baz` the directory `<root>/foo/bar`
 - the one where the top-level package has a stubs
   package masking it, e.g. for `foo.bar.baz` the
   directory `<root>/foo-stubs/bar`

The fix is pretty straightforward:
- modify the relative path finding to return a list
  per search root instead of an optionl, and flip
  the `List.filter_map` to `List.bind`
- tweak the internals to check both possible relative
  paths.

This commit fixes the unit tests in-place so it's
easy to see inline what happened; the resulting code
has duplicate test cases so I'll stack a cleanup
on top.

Reviewed By: grievejia

Differential Revision: D56363471

fbshipit-source-id: f1f2de4129beb5353216f21c710c30baad4c8b66