Deduplicate module tracker tests
Summary:
In the previous commit I fixed a bug in lazy module tracking for PEP 561
stub packages, which resulted in duplicate test cases because the lazy and
nonlazy trackers now agree. I did it that way so that it's easier to see
the effect in the diff view.

This commit keeps the tests actually run identical but deduplicates the
definitions by using the `run_lazy_and_nonlazy` test helper.

Reviewed By: grievejia

Differential Revision: D56363858

fbshipit-source-id: 831d09eb936f49131c1a32b2727e3e2c126a00ae