Allow to model top level functions
Summary:
We consider that the code in the top level of a module or class is actually within an implicit function.
Users might want to skip the analysis of that function if it is slow to analyse. This diff allows this, using a special name `__top_level__` and `__class_top_level__`.

For instance:
```
# test.py
foo(bar()) # slow expression, want to ignore this

class Foo:
  foo(bar())
# taint.pysa
SkipAnalysis
def test.__top_level__(): ...
SkipAnalysis
def test.Foo.__class_top_level__(): ...
```

Reviewed By: alexkassil

Differential Revision: D56362474

fbshipit-source-id: 13f67951e681fe4c208c7d2e425ca5b3929c4988