Push the reference -> reference aliased exports logic into typeCheck.ml
Summary:
I wanted to delay this until after implementing the logic itself because it
can't be directly unit tested once we remove it from the GlobalResolution
interface - the only way to actually execute it is now inside of `typeCheck.ml`
code.

But at this point if we are happy with the logic, we can go ahead and delete
it from the unannotatedGlobalEnvironment and GlobalResolution interfaces,
and instead use a helper function in `typeCheck.ml` so that the ugly
fully-qualified name logic lives in the same place as the only two callsites
that need it.

If anyone feels strongly that we should keep the tests, then the most correct
solution is probably to just keep the function in UGE so that it can be unit tested
easily, and maybe rename both `resolve_exports` and `legacy_resolve_exports`
to have clearer, more descriptive names.

Reviewed By: grievejia

Differential Revision: D56403171

fbshipit-source-id: b9db797bfdc85377a63d89e1ec1c76b78944d804