Re-implement `legacy_resolve_exports` in terms of `resolve_exports`
Summary:
This diff deletes the existing "fixpoint" based logic of
`legacy_resolve_exports` and replaces it with best-effort attempt to use the
much more principled `resolve_exports` when mapping a reference that might be
prefixed by an import alias to the cannonical reference.

This is a first step toward deleting `legacy_resolve_exports`, since we can
now push the logic up to the callsite; I wanted to cut it over first because
once we move the logic the existing test coverage will go away, and it's good
to have in our commit history exactly what behaviors changed.

The cases that changed always involve a collision in fully qualified name;
`legacy_resolve_exports` preferred assuming that the reference is intended
to be the one in the most-deeply-nested submodule whereas `resolve_exports`
explicitly assumes (when `from` is not provided, which it almost never is
in the *current* Pyre codebase - this will change when we remove qualification)
that we should prefer attributes of less-deeply-nested package modules.

We do a prefix check to ensure that this implementation is idempotent, so that
calling it repeatedly cannot diverge (there does exist an edge case, involving name
collisions, where naively calling `resolve_exports` and flattening repeatedly
would diverge into an unbounded number of attribute lookups... the
`from outer.inner import inner` test case illustrates how divergence could occur).

Reviewed By: grievejia

Differential Revision: D56315652

fbshipit-source-id: f3c7375950c5401876e768a8e0c92a215ad7fa7a