Rename type alias functions to be clearer
Summary:
Aliasing has two meanings in Pyre:

- "normal" import aliases and the resulting name aliases of normal
  things like functions, which is what `resolve_exports` and basic
  scope analysis work with. We don't actually use this term a lot
  in our *code*, but it comes up a lot in *discusssions* especially
  around qualification

- type aliases, which are a special case of ^^ and require special
  handling in Pyre (I'm not sure yet if the special handling is
  intrinsic or could be eliminated with better symbol tables; my
  guess is it's actually intrinsic assuming we want to preserve the
  aliased names in error messages and IDE hover results)

The alias environment code refers specifically to the latter notion
of type alias, which is something I didn't realize for a while when
working on Pyre, and it's *still* easy to get (briefly) confused
when I see it out of context like `GlobalResolution.get_alias`.

This change was triggered by me reading the `forward_assignment` code and
getting annoyed that I was briefly confused why `GlobalResolution.get_alias`
was being used to decide whether an assignment is defining a type alias;
the problem was my mind first jumped to the other kind of alias.

Reviewed By: grievejia

Differential Revision: D56516980

fbshipit-source-id: e52ebe8a936c3f5dcd1d3989ea6b90a977c098e3