Rename pyreParser -> pyreMenhirParser
Summary:
We have three parsers in Pyre:
- the old parser, a hand-written Menhir one that we keep because
  pysa needs (a simplified version of) it for the model DSL, and
  Pyre uses in a couple places just because we happen to still have it
- the CPython parser, bound to ocaml in the opam package `pyre-ast`
  and exposed as `PyreCPythonParser`
- the error-recovering Errpy parser that wraps tree-sitter, provided
  by `PyreErrpyParser`

Of these three, the only one with an unclear name is `PyreParser` which
is the old one that doesn't actually support even 3.10 syntax in general
and shouldn't be used for "normal" cases.

This commit renames it so that we aren't calling this legacy parser
the `PyreParser` anymore - it would be reasonable for anyone not familiar
with the codebase to think this is "the pyre parser" which would be super
confusing.

Reviewed By: kinto0

Differential Revision: D56530186

fbshipit-source-id: 4c85a3808e468b5679659c0116ec23243e3c73f6