Add test case documenting assignment bug
Summary:
In the previous diff I wrote a detailed domment describing
`forward_assign`, and I also noticed a bug in multiple assignments.

The bug has to be fixed as part of introducing subscripts to Pyre's
AST, so I want to have a test case exercising the bug before I get
started.

Reviewed By: grievejia

Differential Revision: D56519571

fbshipit-source-id: afa9609649116589679315699fdb5336388d60f6