Do not call preprocessing in Type.ml
Summary:
We currently have a code path to parse raw strings into types,
which for various reasons we cannot delete right now. But it
seems crazy that we would need to preprocess a single type
expression - most preprocessing actions are either about module
level scope information or involve handling special kinds of statements
and such.

But I would very much like to stop parsing an entire "source file"
to do this, and stop running preprocessing on it - preprocessing
is a huge black box, and we *hope* that nothing in it is important
for a single already-fully-qualified type expression.

I skimmed the preprocessing stages and can't find anything that seems
relevant; even the "expand string annotations" step isn't because in
this situation we have no way of knowing the expression is a type.

Reviewed By: grievejia

Differential Revision: D56547123

fbshipit-source-id: b157d871ab846419ae1e250bd2e9f47cdb126357