Document `forward_assign` in some detail
Summary:
The `forward_assign` logic is pretty nontrivial to skim, so I added
a fairly detailed comment of the cases we need to handle at the top
so that you can get oriented before reading the code itself.

In the process I realized we have a pretty significant bug that
as far as I can tell was not tracked - the `forward_assign` logic
ignores subscripted left-hand-sides entirely. In some cases
this is fine because we currently eagerly lower simple assignments
such as `x["k"] = v` into `x.__setitem__("k", v)`, but that hack
isn't composable, and we never validate the same thing inside a
multi-target assignment - for example even if `x["k"] = v` produces
a type error, `x["k"], y = v, 42` will never error.

I included a TODO in the `forward_assignment` description indicating
this, and attached to it the task to simplify AST subscript operations; that's
fine since the reason I was reading the code in the first place is that I'm hoping
to work on removing subscript lowering :)

Reviewed By: yangdanny97

Differential Revision: D56518106

fbshipit-source-id: 3fefa53ad5057301a9231fa1ae5b222c17d0b7b5