Refactor forward_call [3]: delete `arguments` from callgraph builder
Summary:
**Context**

It is not desirable for the callgraph builder to require argument expressions,
because I want to make most of the current `forward_call` logic able to
dispatch synthetic calls where the arguments may only be available as types
(the immediate use case is to model runtime behavior of dispatching to
`__setitem__` in complex assignments such as `x["k"], y =
some_function_call()`; in this context there is no expression that corresponds
to the `__setitem__` argument, but we already have the logic to work out it's
type based on `some_function_call()`.)

**This diff**

The callgraph builder never needs to know the arguments of the callee,
I have no idea why the type signature includes this - maybe historical
reasons, or maybe the author thought that they were likely to need
arguments in the future.

But it's actually wrong to require the arguments. We know we don't need
them (I'm deleting only dead code here), and the `__setitem__` example
establishes that in general the callgraph might include synthetic calls
that don't actually have any arguments as expressions.

Reviewed By: grievejia

Differential Revision: D56601721

fbshipit-source-id: 641579577e608c1e34a2387668fa2a20f7c7e59d