Refactor forward_call [2]: extract `forward_call`
Summary:
**About this stack**

As discussed in more detail on the previous diff, I am working on making it so that
`forward_call` logic can be used outside of `forward_expression`, and can be used
at the type level instead of requiring an expression for the arguments.

This is important so that we can synthesize calls based on type information, which is
needed in order to correctly model the Python runtime - for example (and my immediate
use case), properly type checking multiple-assignments that include index expressions
on the LHS such as `x["k"], y = some_call()` requires the ability to type check a
call to `x.__setitem__`, but using only type information (because we don't actually
have an expression corresponding exactly to the call - what we have is the ability to
understand the iterable behavior of the return type of `some_call()`).

**About this diff**

This commit follows up on the previous one by extracting forward_call into a top-level
mutually recursive function instead of having it be a nested function inside of
`forward_expression`.

This almost makes it usable from `forward_assign`, but there's one more problem I need
to fix first, which is that I need to be able to pass arguments as types rather than
as expressions. I'll work on that next.

Reviewed By: grievejia

Differential Revision: D56601722

fbshipit-source-id: 5a58e542dc01fd478b15ff846d29bba36e9edeb6