Refactor forward_call [4]: Deduplicate resolving arguments
Summary:
**Context**

I'm working on making a version of `forward_call` that can be used
to handle subscript assignments, which requires being able to handle
calls with arguments that exist only at the type level but not as
expressions.

To do that, I need to separate the part of the logic that resolves
argument expressions (which is important, but only when we are
operating at an expression level rather than a type level) from the
rest of the logic.

The nice thing is that this is a nice BE win anway - we were duplicating
a nontrivial chunk of low-level code to do exactly this operation in
3 places.

This diff extracts a helper function to the top of `forward_call`,
which will allow me to split that into a separate function in the
next diff.

Reviewed By: grievejia

Differential Revision: D56601724

fbshipit-source-id: 26f6263ea8f697dbe1e61982babcaef8697e3a0e