Refactor forward_call [1]: Improve BroadcastError handling
Summary:
**About the next several diffs**

Because Pyre's handling of assignments is broken in general right
now due to only validating `__setitem__` in simple assignments, I
want to stop using `__setitem__` rewrites and instead dynamically
create a "synthetic" `__setitem__` call for Pyre to check when
I encounter subscripts in the LHS of an assignment.

Unfortunately this is actually going to force a major refactor,
because `forward_call` is currently nested inside of `forward_expression`
and requires its arguments as expressions rather than types (in other
words, the current codebase can only synthesize a syntax-level call,
not a type-level call, which cannot possibly model python sementics
corectly - we *have* to be able to "dispatch" `__setitem__` at the
type level in general because you may not have an actual expression).

As a result, in this diff I'm kicking off a refactor to extract
the `forward_call` logic and generalize it enough that we can use it
for synthesized `__setitem__` calls. This is a nontrivial refactor that
will take a few diffs.

**About this diff**

This first diff eliminates the only direct use of the outer call
expression from `forward_call` - we use the outer expression in order
to convert `BroadcastError` from a return type into a type error, because
we want the expression to be available in the error message.

We don't need to be able to do this in all cases - only "normal" calls
(including binary arithmetic operators) can actually produce this kind of
error, other cases like `super` or synthetic `__setitem__` don't need
it. As a result, we can pull the logic out of `forward_call`, and
instead only run it in the "normal" case within `forward_expression`.
This feels "right" to me, since the whole point is that we want to
use the expression in the error message.

Note that this is specific to shape type support, which is experimental
and we've considered removing. Nontheless, I want to do the refactor
because (a) I'm not looking to remove shape type support right now, and
(b) I actually think the use case of a return type that becomes a type
error is potentially useful in other contexts, so this code pattern is
actually not necessarily bad.

Reviewed By: grievejia

Differential Revision: D56601725

fbshipit-source-id: 21f36e602448e217d77ee3c38d295b4dccc11135