Refactor forward_call [5]: Split function call handling
Summary:
**Context**

In this stack I've been working to split the logic for type checking
an actual function *call* (which involves handling signature selection,
type errors, and type inference given a callee and the argument type
information) from the checking of a call *expression*, which requires
doing the same for all of the arguments.

The underlying reason is that to handle some special function calls
that happen under the hood in Python - my immediate example is the
implicit call to `__setitem__` when the LHS of an assignment is a
subscript - we sometimes will have to check a call where the arguments
exist only as type information, not actual AST expressions in code.

A snippet motivating this is an assignment like
`x["k"], y = some_call()`, where we can statically analyze it by
using the return type of `some_call()`, but there's no actual expression
directly corresponding to the value we are assigning to `x["k"]`.

**This Diff**

This diff actually makes the split.

For lack of a better name, I called the existing signature that handles
folding over arguments `forward_call_with_arguments`, and the newly
extracted function that handles just the call itself `forward_call`.

Reviewed By: grievejia

Differential Revision: D56601723

fbshipit-source-id: 3d7f18432936d0932761b95c9fe4f8ff2997f814