Extend the test case of the subscript bug.
Summary:
As I started looking into a fix for the subscript handling bug, I realized
that there's a subtly different problem from the one my new test case demonstrates
that also has to be handled correctly: any expression can have an effect on the
current typing resolution (context, in non-Pyre-specific vocabulary) and type errors
so we need to be sure to pass the resolution around correctly.

I thought it would be a good idea for the test case to exercise all of this, so
that I can be sure my fix doesn't mishandle anything (it's actually very easy
to accidentally fix only the problem that the test covered before this diff but
leave this bug in place - `resolve_type` would work just fine for getting the
key type but would drop the resolution and type errors changes).

Reviewed By: grievejia

Differential Revision: D56641453

fbshipit-source-id: 4dbc0064b7b59145dad9f1458b9cbdb0f34002bc