Add an integration test for InitVar attribute table logic
Summary:
We don't want to allow InitVars into dataclass attribute tables, they
are there *only* to indicate parameters to `__init__` and `__post_init__`.

But our unit tests compare attribute tables, which can make it hard
to see a behavior change in unit tests when we fix the problem. Let's
also cover this in an integration test so that we can verify the fix.

Reviewed By: migeed-z

Differential Revision: D56658179

fbshipit-source-id: 849d55f015c821436fdb2819877b975bff8d4485