Add a test illustrating a qualification bug for subscript assignment
Summary:
We currently do not properly qualify subscript assignment targets - we skip over
qualification for expressions that aren't of a couple of forms we recognize, which
does not include subscripts.

This results in bugs in analysis, as I discovered when improving unawaitedAwaitableCheck;
I will fix the bug shortly but I want to add a test case first so that the commit fixing
it can illustrate the fix.

Reviewed By: samwgoldman

Differential Revision: D56706549

fbshipit-source-id: b1a3f2fc1d4dda5ce52fac7784da1cda286eeca4