Handle multi-target assignments in unawaitedAwaitableCheck
Summary:
Up until now, Pyre did not traverse targets of assignments looking for await
expressions, which meant that the use of an await inside of a subscript
target (which can contain arbitrary expressions inside the base or the key)
would produce a false positive.

In addition, the underlying architecture issue (based on the assumption that
the parser can easily rewrite to `__setitem__` calls) is blocking us from
making our parser more compatible with CPython.

This commit adds support for subscript targets (which for now the parser is
expanding to `__getitem__` calls - we will change that to a new subscript
AST node later after the `__setitem__` case has been handled), which
immediately fixes the false positive for subscripts in multi-target assignments
and also will allow us to remove the `__setitem__` rewrite from the parser.

**Remaining bug: qualification**

Notice in the test results that the multi-assignment case for an awaited
subscript *base* is now fixed, but it's not fixed for an awaited subscript
*key*. The logic in this diff actually does handle keys correctly, the
problem is that qualification isn't dealing with subscripts in targets
correctly. I will make that fix separately since it is in a different
part of the codebase, and we'll see changes to the test case then.

Reviewed By: samwgoldman

Differential Revision: D56706547

fbshipit-source-id: 0e52a090124eabd6bd9711f887a3003f54757ab7