Add tests to unawaited awaitable for subscript assignment
Summary:
Add new tests verifying that we do not correctly traverse subscript expressions
in multi-target assignment statements, which leads to false positives.

In addition, the tests demonstrate that we do not seem to traverse the key expression
even in simple subscript assignment statements. This surprises me since the parser
eagerly rewrites simple assignments into `__setitem__` calls. I am not yet sure
the root cause, I filed T187209002 to investigate whether there are general issues
with method calls.

But I also think that it isn't relevant to the subscript change per se; if we
properly handle subscripts in general (including getting rid of the rewrite
to `__setitem__` in the parser) then the false positive in this particular example
should go away on its own. As a result I marked the false positive in the single
assignment case under the same AST fix task.

Reviewed By: samwgoldman

Differential Revision: D56706551

fbshipit-source-id: 7545ecef2893312e550d0add1a46daa0d0d16f6f