Handle stubscript assignment targets in `Scope.ml`
Summary:
Up until this stack, Pyre was assuming that all subscript assignments were
rewritten by the parser into `__setitem__` calls, which are a completely
different kind of statement.

But there are two problems:
- we want to unify our AST with Python's by not doing that rewrite, and
- the rewrite doesn't actually work for multi-assignment targets such
  as `x[k], y = v, 42`, so Pyre has a lot of bugs related to not traversing
  the `x[k]` correctly.

This diff is part of a stack where I fix the multi-assignment handling, and
in the process prepare Pyre's analysis code for a parser change where we
get rid of the `__setitem__` rewrite (at first, we'll have a `__getitem__`
rewrite instead, then as a follow-up I'll create a proper `Subscript` expression
we can use).

Here, I'm fixing scope traversal because the `x` and the `k` above can be any
expression, not just simple names: it's perfectly legal python to write things
like `function_that_returns_a_dict()[key := k] = v`, which means that the
scope can be impacted by walrus operators inside of either the subscript base
or key.

Reviewed By: samwgoldman

Differential Revision: D56709471

fbshipit-source-id: d0e9a31762a5e643a5ca34e4845eb7f3f4a3029c