Remove unused spawn-per-job worker support
Summary:
Under Linux and Darwin, workers are pre-spawned, listen for jobs on a channel,
and fork for each job.

A long time ago, Flow added support for Windows, which does not have a `fork`
implementation. To support Windows, Flow would spawn a process for each job,
that process would run the job, then exit.

Pyre does not currently support Windows, so we can remove this code without any
regret. Furthermore, the performance of spawning a process per job is terrible.
Under Windows, spawning a process can be quite expensive, so Flow abandoned this
model anyway in favor of long-lived workers.

Reviewed By: grievejia

Differential Revision: D56716548

fbshipit-source-id: 9573fa276b212d4e4a8471f8236fdafab0038419