Add a test of subscript assignment FN in uninitializedLocalCheck
Summary:
There is one more bug I found when working on subscript assignment targets,
which in the simple case Pyre's parser rewrites to `__setitem__` calls (we
want to stop doing this) but in the multi-target case Pyre failed to analyze
at all until the current stack.

The bug is that uninitializedLocalCheck (which uses a separate AST traversal from
the ones we've already fixed in typeCheck and unawaitedAwaitableCheck) is also
ignoring subscript assignment targets.

This commit adds a test illustrating existing behavior, the next one will fix
it and we'll be able to see the behavior change.

Reviewed By: samwgoldman

Differential Revision: D56709472

fbshipit-source-id: 888248ac07a4339ec1982b28db6075691b29a7db