Validate subscript assignment target in uninitializedLocalCheck
Summary:
This commit fixes a false negative where Pyre was not taversing the base and
key sub-expressions of subscript assignment targets in `uninitializedLocalCheck`.

As far as I know, this is the last of the bugs caused by not handling this case.

**Stack Summary for subscript assignment target analysis changes**

Because this checklist could be useful for new control-flow changes (either fixes
to our AST or new syntax), here's a list of all the places I changed and where
you can most easily test it:
- Scope.ml: most easily verified using false positives in uninitializedLocalCheck
- Qualification inside Preprocessing.ml - not handling this tends to lead to
  weird behaviors elsewhere, sometimes false positives and sometimes false negative;
  the errors will tend to look like missing symbols in the symbol table.
- uninitializedLocalCheck.ml - you only need to fix false *negatives* here, because
  the initialized locals are built by Scope.ml code and hence false *positives*
  are fixed there
- unawaitedAwaitableCheck.ml - you need to traverse subexpressions; the actual logic
  changes should be simple as long as the new syntax doesn't bind new names (this
  example doesn't - we just need to traverse more expressions, but not bind them)
- typeCheck.ml - this one is usually pretty obvious and well-tested. The tricky
  thing is remembering all the *other* places above we have to update.

I do wonder if there's a way to share more code between some of these

Reviewed By: samwgoldman

Differential Revision: D56709474

fbshipit-source-id: 530fa142c280e8c16839afa681a1677b8b38328b