Clarify code to read worker response
Summary:
Pyre uses a pair of pipes to send jobs to workers and read the results back.
When we read results from the worker, we check to see if the worker has exited,
so we can raise the correct error.

Before this diff, we first checked the worker process with `waitpid` with
`WNOHANG`, which returns `0` if the process is still alive. Then, we do a
blocking read on the pipe.

This diff changes the order. First, we try to read the response from the pipe.
If that fails, then we check to see if the worker has exited using `waitpid`.

Originally, I wanted to make this change for clarity. Primarily, reading a
response is waiting to read a value from the pipe. Moving this read to the
beginning of the function makes this code easier for me to understand.

However, I realized this code also fixes a "time-of-check to time-of-use" race
condition in the code. The worker could exit after we waitpid but before we try
to read from the pipe. In this case, before this change, the read would fail
with a broken pipe exception, but would not be logged as a worker exit.

By reading from the pipe first and then checking the process status with
waitpid, we are guaranteed to detect a worker exit that leads to a broken pipe.

Reviewed By: stroxler

Differential Revision: D56735008

fbshipit-source-id: a9acafde8a3857363fef95246647809456c321a8