Bump the default python version for tests to 3.12.2
Summary:
Test behavior can be confusing when working with recently-added typing features,
which are always gated behind version flags, if the default python version is
lower than the one we are trying to support.

Note that in production, we're often going to want a *lower* default
version, because the tests are helping us build for the next version we need
to support whereas the default in prod should be whatever we fully support.

This diff decouples the unit tests' default version (which is now defined in
`ScratchProject.setup` from the command line entrypoints' default version
(which is defined in `configuration.ml`).

No tests break as a result of the bump.

Reviewed By: yangdanny97

Differential Revision: D56761458

fbshipit-source-id: d70acab96d51577e34509442a497c6b4a1c58b5b