Don't use `|` operator in dataclasses_json code yet
Summary:
Using a runtime `|` breaks pyre on 3.8, which blocks open-source release:
https://github.com/facebook/pyre-check/actions/runs/8899961003/job/24440430168

But using `from __future__ import annotations` breaks dataclasses_json, which
assumes the types are proper runtime data.

For now we have to use `Union`; once 3.9 is past end of life we can use ` | `
freely (it does work with dataclasses_json on 3.10+).

Reviewed By: samwgoldman

Differential Revision: D56788805

fbshipit-source-id: c2ffaf7e95405beac19c4fda2c9901858325f9c8