Fix bug propagating errors
Summary:
Our type check interfaces are not very consistent - they always
pass around `resolution` as state, but only sometimes type errors.

As a result, you have to be careful about when we are using
writer-monad-style semantics where each function just appends to errors
versus when the errors from various subcomponents of a complex
statement or expression have to be concatenated explicitly.

The bug here was that I failed to explicitly concatenate errors from
the right-hand-side (which can't be passed to `forward_expression` as
state) with the errors coming from traversing the argument and call
(which do expect to tack errors onto a preexisting list).

It would be nice to think much more carefully about how to get the best
consistent structure for this kind of thing, it's pretty error-prone to have
the error passing style be inconsistent.

Reviewed By: samwgoldman

Differential Revision: D56798431

fbshipit-source-id: bbd4a2c35e5758e39d4a41f1650c752bf7d373c2